optimize and modularize
